Rename Solarium endpoint to clarify purpose
We will be configuring Solarium through our `.env` file, not through
the provider config. Additionally, our current SolrScoutEngine follows
the Algolia example in that it uses `searchableAs`, which assumes that
each model will have it's own index (i.e. core?). I'm uncertain whether
that's how Solr is meant to work, but we intend to store all models
under a common schema anyway, so this is the first step to moving
in that direction.

https://solarium.readthedocs.io/en/stable/getting-started/#basic-usage