Make all desired models searchable
This is an unstable commit. AFAICT, everything is stable on the
Laravel side. However, there will need to be some reconciliation
in the Solr schema. I am currently running into an issue where
the GUID id's are conflicting with the numeric id's – because
the numeric id's are indexed first, and because I'm currently
using a managed schema, Solr thinks that the id field should
be an integer. More work is needed to define a stable
`uniqueKey` field for Solr.