Allow passing preference to search for order stability
Without this, items will jump up and down on repeated searches.
AFAICT, this is due to results coming from different shards, maybe
even in a different order. Passing `preference` should stop that
randomization. Note that `preference` cannot be a constant, since
that will cause uneven load balancing. Ask the client to e.g. pass
the session id for the current user as the `preference`. We could
do this server-side, but I'd like to avoid saving state for now.