Finish refactoring "subresource" controllers
The diff for this commit looks crazy, but the changes are mostly
simple. Look to `ApiNewController` for the architecture, and to
`ArtworksController` for examples of how it's meant to be used.
Route definitions have changed to reflect the new actions. No
external behavior has changed, so Swagger has not been updated.