Differentiate b/w successful and attempted runs
There is no `new Carbon('never')` equivalent, so we needed to come
up with an arbitrary value for `never` that would function well for
purposes of data comparison. Using a Unix timestamp of 1 is as good
as anything: not abstract, but practical.

Here, we are wrapping Laravel's `Command::execute` into our own
implementation. This should allow all `handle` methods to function
as usual, while still allowing us to do logging.