Remove VerifyCsrfToken middleware
I think we can remove this for now, since none of our routes trigger
anything that writes to a database. We can revisit this later.

Plus, we aren't using the `web` route group currently, so it wasn't
being triggered. From that perspective, everything relating to the
`web` route group could be removed without breaking the app.