Delete all front-end files
Our application will have no (HTML-based) web user interface, or rather
the `data-dashboard` repo serves as the de facto interface. Separating
the dashboard codebase requires us to work with the API directly, which
is a good thing, as it forces us to interrogate its capabilities.

Anyway, we don't need these files for the foreseeable future.