Move Membership into module
...except tests. Those will need to be handled later, and will involve
modifications to `phpunit.xml` to detect tests in module subfolders.

I think it's useful to see which files outside the modules needed to be
modified during this process. This indicates tight coupling that should
be addressed. Specifically, `Search` appears to be tightly coupled to
all potential modules.

Still debating where modules should live:

 1. /modules in base_path as is the case now
 2. /app/modules
 3. /apps and move everything that remains in app into a `Core` module