Refactor tests into a declarative approach
This approach is quite similar to that undertaken for the previous
controller refactor. We abstracted all the `fetch` and "HTTP code"
tests into our base `TestCase`, and declared the desired model and
route for each test. All of the test pass.

This was done to reduce code duplication, with the aim that future
tests will build off this framework: we can write them once inside
`ApiTestCase` (or a trait used by it), and inheritance will ensure
that they run for each resource.

I intend to implement this approach for testing that the `fields`
param works correctly.

The next step will be to refactor the `404s` tests. I left them for
another commit, since they require special logic for generating ids,
which tends to be specific to the model, or the system of record.