Remove footnotes from the data aggregator
This is a complex topic, but overall, normalizing footnotes into their
own models would (currently) complicate import and search, yet offer
marginal benefits.

Firstly, linking to footnotes is currently impossible. For example:

https://publications.artic.edu/ensor/reader/temptationstanthony/section/292#fn-292-10

One might expect this link to open footnote #11 (not #10), which is in
paragraph #8 of that section. However, when that link is visited, the
reader will *not* jump to the correct paragraph, and the footnotes pane
will *not* open to show the correct footnote, as it does if a footnote
is clicked directly. Fixing this error would require modifying the
source system, which we would like to avoid.

(This is however a minor point. The following are more pertinent.)

Secondly, footnotes contain no significant data besides their contents.
(That is, within the source system, which is a modified OSCI Toolkit.)

One might envision that at some point, we might be able to parse these
footnotes to derive bibliographic identifiers for the cited sources,
perhaps generating links to an external authority. Then, it might be
worthwhile normalizing footnotes into their own model. Currently, such
links are not provided by the source system, and it is highly unlikely
that we will be able to undertake this project within the near future.

It should also be noted that footnotes do not contain true identifiers,
or rather, their identifiers are somewhat misleading. Their ids are
based on their order within their section, which might be acceptible,
but there appears to be an off-by-one-error present, as illustrated by
the link above. Footnote ids are zero-based, while footnote rendering
is 1-based. Until that error is resolved, we should be vary of relying
on these identifiers.

Lastly, parsing footnotes into their own model would complicate import
and search configuration. For a number of reasons, we will have to rely
on the OSCI Toolkit's "EPUB API" to obtain the data we need. This API
delivers content on a section-by-section basis. As such, footnotes are
already baked into the primary "unit" of data we are recieving from the
source system. Parsing them into a separate model requires extra steps.
That in itself is not a roadblock if the cost-benefit was worth it, but
given the current state, this does not yet seem to be the case.

Prior to this commit, footnotes were not marked as searchable. This is
a good approach, since their presence would skew search results for the
site: among other factors, search relevance is determined by the length
of the field in which a search term appears. The logic behind that is
that this makes it more likely that the document is *about* that search
subject. Footnotes are short by nature; from a UI perspective, they
would likely be grouped with other "DSC" search results in any site-
wide search interface. Their brevity would boost them above "section"
results, which is definitely undesirable.

Furthermore, footnotes constitute a O2O relationship with sections, and
their contents are already included in section text output (EPUB API),
so it doesn't make much sense to parse them out for separate searching.

This topic should be revisited if we do attempt to correlate footnotes
with their bibliographic identifiers. This will not be a minor effort.