Remove DSC TitlePage
Title pages are a valid content type within the OSCI Toolkit, but for
the time being, we will remove them from the data hub. Their syntax
is malformed, and their content is limited to cover art + publication
title. This should be revisited when a more holistic export attempt is
underway for DSC content.

Note that a title page can contain multiple figures.