Disable relationship b/w Assets and Agents
This is a temporary fix.

This was preventing the image import from running, since it was trying
to create duplicate artists. I think the required column is missing
from one of the tables.

Before fixing, it might be a good idea to combine the various asset
types into one table, and spin out the interpretive resource fields
into another. Needs more thought.

All tests pass.