Use accessor properties to get related ids
One of the cool things about Eloquent is that it will return `null` if
it cannot find an associated resource. This allows us to use `sync` w/
relative impunity, while operating with the assurance that the clients
of our API will not receive results that include id references to non-
existent objects.

The trick is that we must go through the association to take advantage
of this functionality. If we want an id, we need to at least check if
the assoc. object exists first, rather than returning the column that
contains the id.