Add import:dsc command
Because the OSCI Toolkit offers no reliable metadata re: last modified
dates, it will be necessary to re-import the full set of catalogues
every time. This shouldn't be a problem: each doc is fairly hefty,
since each section is a book chapter, but there aren't too many of
them to make it unmanageable.