Move seedFoobar methods from Artwork to seeders
I think having them with the Artwork was useful when we were making
seed calls during our import process, but now that we've separated
the seeding process from the import, it makes sense to move these
methods to their respective classes.

(Likewise for the previous commit re: Exhibitions btw.)

Our refactor/faker efforts allowed this to be a smooth transition.

All tests pass and the seeders run successfully.
Next steps are to shorten and consolidate them.