Add seedPivot method for BelongsToMany
This reduces the number of manual loops we need to write.

Eloquent's BelongsToMany is the only relationship that uses a pivot
table. I'm unsure as to the naming of this method, but I hope that
we can implement others for BelongsTo and HasOne, and allow calling
them dynamically through a generic `seedRelation`, which delegates
to the correct submethod by type-checking the return of `$method`.