Change semantics of seedBelongsToMany
I'm trying to change the semantics from parent-child to a more
subject-object approach. This is more transferable to future
`seedRelation` methods.

The basic idea is that a subject should alway have a method that
can be used to attach objects, and an object is always the class
that gets attached.