Refactor search request logic for non-text queries
Appears stable. The `q` param is no longer required, and the `query`
param is now the deciding factor between whether or not the thin client
should act as a pass-through. Our custom boost logic is still applied
to searches made using `query`, which seems acceptable for now.

If `q` is passed, search suggestions will be provided. If it's passed
alongside `query`, its only purpose will be for search suggestions.