Use custom fork of tamayo/laravel-scout-elastic
This will treat your `ELASTICSEARCH_INDEX` env as a prefix for Scout.

So as to avoid breaking the thin search client, we needed to add an
`ELASTICSEARCH_ALIAS` envar. This way, we can target all indexes for
simultaneous search. For per-resource search, we mirrored the index
naming logic used in our fork: when searching a resource endpoint,
`ELASTICSEARCH_INDEX` will be prepended to the resource name to
determine which index to search.

We will need to update alias-related commands to take advantage of the
new envar. We'll need to rename the index envar to clarify its purpose,
and update references to it across our codebase.

Necessary for one-model-per-index support until the related updates
are incorporated into a release in the main repository. See commits:

https://github.com/art-institute-of-chicago/laravel-scout-elastic/commits/index-per-model

Based on code by @hoersten submitted in this PR:
https://github.com/ErickTamayo/laravel-scout-elastic/pull/62

Elasticsearch 6.0 compatibility discussion:
https://github.com/ErickTamayo/laravel-scout-elastic/issues/90

Composer settings adapted from here:
https://stackoverflow.com/questions/21314381