Remove fallbacks for alias and index envars
It looks like all the various fallbacks we've written have gotten out
of sync over time. For easing maintenance going forward, I've removed
all fallbacks for these envars. This isn't something that we want to
recover from gracefully. If someone enables the `SCOUT_DRIVER` in their
`.env` without providing `ELASTICSEARCH_INDEX` and `_ALIAS` envars,
that's something they'll have to fix on their end.

Going forward, we should avoid providing fallbacks for any envars that
get used often throughout the codebase. If an envar is invoked once or
twice, e.g. in a config file, it's fine to use fallbacks. Otherwise,
it becomes harder to maintain, e.g. when naming conventions change.

As an alternative, we could use `config()` instead of `env()` in these
instances. This will allow us to provide a fallback in the config file.