Remove CopyrightRepresentative model
"Copyright Representative" is not a valid AgentType. The status of
being a copyright representative is bestowed on an agent when that
agent is related to an artwork through its copyright representative
field. It's pretty situational. We'll make it a scope on Agent, not
its own model.

This commit temporarily breaks the distinction between artists and
copyright representatives, but we'll fix that before merging.