Use pivot for artists and copyright representatives
We must make the distinction between "Agent Types" and "Agent Roles".
The first is something an agent is intristically - e.g. a museum, a
school, an individual. The second is something that's defined by an
agent's relationship with another resource. What is an agent to an
artwork? An agent could be the creator (artist) of an artwork, its
copyright representative, or even both at the same time.

In other words, an agent's role is a property of its relationship to
another model. Currently, this is only relevant to agent-artwork
relationships. To represent this, we've added a column to their pivot
table to indicate the `type` of relationship represented by each row.

We use model int constants for the `type` values because they are said
to be slightly faster than using strings. The difference might be
negligible after accounting for indexing. In either case, using model
constants introduces a measure of consistency.

For clarity, we might rename the `type` column to `role` to avoid
confusion with agent types.

Note that an agent could be related to the same artwork multiple times:
once as an artist, and a second time, as a copyright representative.
We'll need to double-check that our syncing logic is solid, and that
we don't detach artists when syncing copyright representatives.

Within the models, we've added relationship methods to retrieve all
agents related to an artwork (and vice versa), as well as relationship
methods to further filter by the type of the relationship. This allows
us to narrow down related agents to artists, or copyright reps.

We've also introduced the possibility of defining scopes on the Agent
model to allow retrieving only artists, i.e. all Agents that have an
artist-type relationship to some artwork.