Refactor Agent routes
Added ability to invoke scope in controllers by route name. Care should
be taken to not bind these actions to routes with non-scope names. For
safety, there's a validation step in place to ensure that the segment
in question corresponds to a valid scope on the controller's model.

Everything works except `/venues` because we haven't yet defined the
`Agent::scopeVenues()` method.