Fix include params to be snake case
This is a Fractal feature. The trick is to specify `$availableIncludes`
in snake case (i.e. with underscores). They will be output via the API
in the same case as they are specified in-code.

We stack some string transformations around the `$_GET['include']`
parameter, which allows us to accept snake, kebab, and camel cases.
`snake_case` can only convert camel-cased strings.

Unintentionally, this is backwards-compatible, though we should
encourage the use of snake case. Looks like the Documentable trait
will pick up on these changes automatically.