Add Agent::scopeVenues
This allows us to list agents that have hosted an exhibition that's in
our database. The downside is that we can't list agents that *can* host
exhibitions potentially without hardcoding ids. Currently, we are only
interested in deriving roles from known relationships.

The `/venues` endpoint works again, since the scope is now defined.