Use two separate tables for artwork agents
So that pivot type thing was a bad idea. We ran into an issue where
attaching a model instance to a relationship derived via the pivot
value wouldn't actually set `type` to the expected value. Our tests
were failing - agents would be attached to artworks, but without the
correct `type` set, so they wouldn't be listed under the right fields.

Now, this is something that we could work around by manually passing
attributes each time we called `attach` or `sync` on the relevant
relationships. We could also try to automate it via this package:

https://github.com/fico7489/laravel-pivot

...but like, this is getting into some complicated territory. The only
advantage we'd have is that it'd be easier to gather up agents which
are either artists *or* copyright representatives of an artwork, and
that isn't one of our requirements. Now that we've identified the
simplest path forward, everything just fell into place. That work
on scopes is still quite useful, going forward. All tests pass.