Copy code from data-service-library
This is a temporary measure. Later, we will split this business logic
into its own repo, and require it both in the `data-aggregator` and in
the `data-service-library` projects.

This is a direct copy, and it probably won't work correctly. This is
necessary from the perspective that it's good to see what changes are
required to reuse the same code in both places. Thus far, only the
namespace references have been changed. We also needed to namespace
the routes, since there's a chance they may conflict with collection
terms and materials.