Add stub docs for library resources
Docs have been regenerated, but the resource descriptions could be
improved. This commit revealed difficulties in targeting namespaced
models in Documentable using typical conventions, which assume a
flat structure. It'd be good to handle endpoint-controller-model
correlations via a `ResourceServiceProvider` of sorts - but one
distinct from the one provided by the foundation. Just some way
to describe these relationships declaratively in a centralized
place.