Rename ArchivalImage to ArchiveImage
This aligns the grammar of `ArchiveImage` with other resources that
require namespacing to avoid collisions with the "primary" resources
in `Collections`, e.g. `MobileArtwork` and `ShopCategory`. This whole
namespacing issue needs a more holistic resolution, but more thought
is required to come up with an elegant approach.

For now, renaming `ArchivalImages` to `ArchiveImages` aligns its name
with that of its data service, which seems like a good step towards
future improvements.

Note that two things were left unchanged on purpose:

 1. The `import:archive` command still targets the `archival-images`
    endpoint on the data service. This can be addressed if and when
    the data service is updated to match.

 2. The table used by `ArchiveImage` is named `archival_images`.
    This was done to ease this change. We can write a migration to
    update this table later.