Add db:reset command
Useful for resetting a database in lieu of rolling back migrations.
This can be necessary if existing migrations are modified, especially
if tables targeted in the migrations are renamed. Doing so can result
in "orphan" tables that remain in the database. This is just a handy
command to wipe the slate without needing to recreate the database.