Add migration to make all CITI ID fields signed
Apparently, some CITI IDs are negative. It seems that CITI started off
using signed IDs, but at some point a decision was made to expand the
range by starting at the lower bound of the number range, rather than
at zero as originally. Surprising that the schema wasn't modified to
be unsigned instead, but hey, so it is.

This modifies any field ending with `citi_id` to be signed. This should
also catch any pseudo-foreign-key field like `agent_citi_id`.

Also +1 to the decision to call the column `citi_id` instead of `id`.
That choice made this process so much easier. Good call.