Make Elasticsearch field definitions more flexible
See PHPDocs in this commit for examples and justification for these
changes. In short, we need this to create room in our transformation
definitions for other Elasticsearch field mapping params besides just
`type` – and for any app-specific search-related params we may need.

For example, next, we'll use this space to identify which fields should
be targeted when the user performs a simple search (instead of `_all`).