Rename type to resource in search routing
We are in a weird in-between place, where `type` is not an accurate
abstraction anymore. For one, ES 6.0 makes all indexes single-type.
However, our AWS ES instance is currently on ES 5.3, and ES doesn't
support locking down indexes to single-type explicitly until ES 5.6,
so we need to work around this shortcoming by adding a type named
after the index to each index. Additionally, there isn't a 1-to-1
correspondence between models, indexes, and our API resources.

We need to rework our search API to expect an array of resources,
and behind the scenes, translate that to indexes and types.