Add ResourceServiceProvider
This is meant to be the central place for tracking relationships
between endpoints, controllers, models, transformers, etc. It may
aid in API versioning down the line.

We should consider renaming the ResourceServiceProvider in the
foundation to FractalServiceProvider. This could be renamed to
EndpointServiceProvider, but I'll hold off on that for the moment:
there's a chance we might use it to track relationships between
resources later.