Add stub BelongsToOne via pivot behavior
This is a precursor to auto-generating `include` definitions, and an
effort to standardize our relationships. Basically, we have several
many-to-many relationships where one (and only one) from each set
is marked as `preferred` via a boolean column on the pivot table.
The others are considered "alternative" (`alt`). We need a way to
elegantly access both the preferred (sans-prefix) and alternative
subsets via model methods. The ideal scenario is as follows:

 1. Have method that returns both preferred and non-preferred items.
 2. Have method that returns preferred items.
 3. Have method that returns alternative items.
 4. All above methods should return an instance of Relation, i.e.

    Illuminate\Database\Eloquent\Relations\Relation

 5. All above methods should be accessible as an attribute, and
    correctly return only one item for `preferred=true`.

It seems that the "Laravel way" to do this is to extend the `Relation`
class (or one of its descendants) and over-write its `getResults()`
method to call `first()` when appropriate. When a relationship is
accessed as an attribute, Laravel calls its `getResults()` method.

In this implementation, I extended the `BelongsToMany` class with
methods that allow us to flag whether `getResult()` should return an
instance or a collection of instances. This allows us to define a core
relationship, and use it within other methods that returns just the
preferred or alternative subsets: e.g. `images`, `image`, `altImages`.
We further overwrite `newBelongsToMany()` on the base model class to
return an instance of our extended class, instead of `BelongsToMany`.
This should be transparent to all implementations.

If this is a viable strategy, we should consider moving some of this
logic to a trait, or to the foundation, or both.