Move id validation logic to the model
We should update to the newest version of the foundation. I've recently
added a similar commit there, which would allow us to remove the static
`validateId` calls from `AssetsController` etc. See this commit:

https://github.com/art-institute-of-chicago/data-hub-foundation/commit/fdb624fd05a078ea8455885b5acbf3edb45dec3a

...I'm still trying to remember why these should live in the model, and
not the controller, but I know there was a practical reason for it, not
just good architecture. I think it's because this logic is duplicated
in tests right now. I'll take a look around.