Remove deprecated `type` param from search
As of Elasticsearch 6.0, multi-type indexes are deprecated. Mapping
types will be completely removed in Elasticsearch 7.0. Given this
change, plus the fact that our API resources don't map 1-to-1 to
indexes or types in our Elasticsearch index, we decided to remove
the `type` param in favor of `resources`. We deprecated it in 0.7,
and are removing it entirely in 0.8.

https://www.elastic.co/guide/en/elasticsearch/reference/6.x/removal-of-types.html