Remove deprecated `_source` param from search
We announced its deprecation in v0.7 and are removing it in v0.8.
The idea behind this change is that the user shouldn't care about
the implementation - i.e. about how we store these fields in our
index – they only need to care about the API's output.

https://www.elastic.co/guide/en/elasticsearch/reference/current/search-request-source-filtering.html
https://www.elastic.co/guide/en/elasticsearch/reference/current/search-request-stored-fields.html
https://www.elastic.co/guide/en/elasticsearch/reference/current/search-request-docvalue-fields.html

Plus, we might one day retrieve only the id from the search index,
and grab info from our database instead, passing it through our
transformers. Storage mechanism should be transparent to the user.