Remove explicit text type mappings from some title fields
This may seem backwards. Here, we're taking advantage of ES's dynamic
mapping. By default, when it indexes a string field (non-numeric, not
a date) that wasn't explicitly defined in index-time mapping, it'll
add the field as `text`, plus it'll add a subfield (?) with the type
`keyword`:

https://www.elastic.co/guide/en/elasticsearch/reference/current/dynamic-field-mapping.html

We can't use `text` fields for stuff like aggregations, script queries,
or sorting. But if we've got these `keyword` subfields being generated
for us automatically, we just need to target them when appropriate, and
we're good to go:

https://www.elastic.co/guide/en/elasticsearch/reference/current/fielddata.html

It would be too verbose to define each mapping manually. But we can set
templates for how ES should dynamically map fields, if we want:

https://www.elastic.co/guide/en/elasticsearch/reference/current/dynamic-templates.html

Now, it looks like the designs demand that we create separate `_title`
fields for each term type under Artwork: classifications, subjects,
styles, and mediums - plus maybe their alt variants? Or maybe both?

This can't be sustainable without managing resource relationships.