Add temp `suggest_autocomplete_boosted`
This is a multi-step process. The ideal situation would be to re-index
all docs in search from the database, but that's too time-consuming to
test this new functionality. Instead, we'll add this new field, target
it for autocomplete, create new indexes with this new mapping, re-index
from the old indexes to the new, then import only the boosted artworks
and artists into the new set of indexes. That should make this work.

Later, when we have more time, we can remove the boosted autocomplete
field, and move that functionality into the main autocomplete field.
But again, we'll need to re-index everything from the database.