Retrieve title of source docs in autocomplete
Alright, that didn't work. Looks like `output` was removed in ES 5.0,
which made suggesters more document-oriented. We'll use `_source` to
retrieve the titles instead. My only worry is that we might get dupe
results if multiple name variants match for a given doc. More info:

https://www.elastic.co/guide/en/elasticsearch/reference/5.3/breaking_50_suggester.html
https://github.com/elastic/elasticsearch/issues/22912

If this turns out to be an issue, ES 6.1 added `skip_duplicates`:

https://www.elastic.co/guide/en/elasticsearch/reference/6.1/search-suggesters-completion.html#skip_duplicates