Make route and model in unit tests generic and reusable.
This change creates two new methods for unit tests: route() uses the ResourceServiceProvider to determine the endpoint path to use for a given model instead of requiring each unit test class to define it explicitly. And model() to allow unit tests to provide some logic to determine which model to use for a given suite of tests.