Fix indexing issue with null suggestion fields
Got a `mapper_parsing_exception` with the following reason:
completion field [suggest_autocomplete] does not support null values

With there was a better way to debug Scout, somehow. It doesn't error
out if the indexing is unsuccessful. To see this error, I had to run
`toSearchableArray` and send it manually via Postman.