Use DefaultConfigServiceProvider from foundation
It merges files from `vendor/aic/data-hub-foundation/config/default/`
with those in `config/`. In (most) cases, this allows us to leave out
the files entirely – assuming we modify our settings via `.env` i/o
via `config/*.php` files. Alternatively, we can add these files to
the config folder, but only specify the keys we want to change from
their defaults.

There's a few exceptions to this. For instance, `app.env` is checked
before the "DCSP" has a chance to kick in, and it seems that Horizon
scans for `database.redis` before "DCSP" too. Just something to watch
for when installing new packages.

This may seem overkill within the context of a single repo, but it
does a lot to reduce code duplication across our 7+ microservices.