Remove `image_iiif_url` and `alt_image_iiif_urls`
I've been itching to remove these for a while. We should limit the
number of fields that are passed through from one model to another.

The major exception to that should be when we are trying to derive
distant relations by going through near realtions – e.g. if we're
trying to determine what artists are related to a mobile tour, we'd
have to do something like this:

    Tour > Tour Stops > Artworks > Artists

...but even then, we're only seeking to gather ids, titles, and
includes. Other fields just tend to complicate things.

The `iiif_url` fields are particularly nasty because if the base
IIIF URL changes, everything with a `iiif_url` field has to be
reindexed. Granted, this is unlikely to happen, but having a SSOT
for that data would be ideal.

I'd like to implement a `config` param alongside `pagination` w/
some critical data like the base IIIF URL.

BTW, I confirmed that A17 isn't using these fields. They use a
hardcoded base URL in `app/Libraries/LakeviewImageService.php`.