Rename `medium` to `medium_display`
We found out relatively recently that "medium" is actually a term type.
Like, there's `medium_display` that's a free-text field similar to
`artist_display`, but there's also going to be a `medium` relation
on `Artwork` to `Term` of `TermType` "medium". Renaming the field
is meant to make room for that change.