Add `document_ids` to Artwork
This should fix documents making their way into `alt_image_ids`, but it
revealed a new problem: it seems that documents don't get detached..?
We'll need to test this further. I can see this becoming an issue if
a rep is switched to be a doc, or vice versa.

Stepping back for a min, this could be the case for all `sync` calls.