Expand ArtworkArtist into a pivot resource
Not too sure if this is the right way to do it, but it's a start.

This allows us to run queries like this:

http://data-aggregator.test/api/v1/artworks/111628?include=artist-pivots
http://data-aggregator.test/api/v1/artworks/111628?include=artist-pivots,artist-pivots.artist

I feel like this is going to get strange, but I'm liking the `pivot`
convention for models like this. Outbound transformers deserve a
refactor, but inbound should be completed first.

Still need to add tests and imports for these.