Do detach on relation sync to avoid dupe pivots
I think this was a misunderstanding on my part back then – we might've
been looping through each id in the relation and trying to `sync` them
one-by-one, which would cause a de-sync of previous entries, but if so,
we were doing it wrong, and have fixed it since then. As long as we
sync *all* ids for a specific relation at once, we definitely want
to detach existing relations to prevent dupes and stale data.