Fix `artwork_term` import
It broke because `artworks` were imported before `terms`: `Term::find`
would *always* return false, if the import was done from a clean slate.

I think there is no need to enforce referential integrity on import.
We can do so right before output. If an artwork's pivot table refers
to something that doesn't exist, our `?? null` fallbacks will catch it.

I'm not concerned about performance w/r/t output, because ultimately,
we will need to put some sort of caching in front of our API anyway.
That might be memcached, or an Elasticsearch passthrough. TBD.

In either case, I'm currently more concerned with speeding up imports.
There's a lot of QA that needs to be done in that regard, and running
into bugs on this end is a considerable setback in terms of time.