Disable `links` import
Long story short, `links` isn't an actual data type in LAKE.
It used to be its own data type in CITI, but now any asset
can link to one or more external URLs.

I had an import conflict happen when an asset was imported
as a `link`, and then again, as a `text`. Because the two
are separate models, `findOrNew` didn't trip the `find`
part, and I got a dupe primary key error on save.