Disable 5-min imports on non-prod envs
The problem here is that on e.g. our testing server, both the 11 PM
import *and* the 5-min imports would be queued and executed at 11 PM.
There's no reason to stress the servers – just one run is enough.

Conversely, we do want to keep the 11 PM import on prod, alongside
the 5-min imports – just in case.