Remove `Link` resource
We discovered that "links" aren't a dedicated type upstream anymore.
They used to be in CITI before resources were migrated to LAKE, but
that has since changed. Now, any asset can have one or more external
links attached. However, we aren't actually getting that data yet –
what we are getting is infact a legacy field that isn't directly
editable in LAKE. This will need to be revisited later.