Install xethron/migrations-generator
Useful for populating `down` methods in migrations. Might be useful
for squashing existing migrations, if and when we decide to do so.

Try running this command to test it out:

    php artisan migrate:generate terms,categories

Answer `n` to the prompt. Answering `y` marks those migrations as
having been run – the reasoning being that these tables already
exist in the database. That's irrelevant to us if we're just
copy-pasting from these migrations.