Remove `artworks/{id}/terms` route
This isn't terribly useful, given that it's lossy - there's no way
to distinguish which terms are preferred - and if we are to combine
terms and categories, I don't want it getting in the way.

Overall, I'd like to have these sub-resource routes auto-generate.
Each relation should be eligible for a sub-resource route, without
any boilerplate code on our part. Personally, I'm not a fan of
sub-resource routes in general, but if we are to have them, I'd
like them to be implemented consistently across the board.