Remove TermType resource
This is a step towards combining categories and terms into one model.
The reasoning behind this removal is as follows:

 - TermTypes are specific to Terms, they don't apply to Categories

 - If we want to roll additional "keywords" or "tags" that don't come
   from our collection system into the CategoryTerms model, TermTypes
   would become even more irrelevant.

 - TermTypes are very static. We tightly control which TermTypes are
   imported in `data-service-collections` – we go so far as to list
   out which ids we want to import! Having these change is rare.

 - We *treat* TermTypes as static values. Note how TermTypes had a
   list of public static variables, referring to specific ids, and
   how Terms drew on that list, rather than the database, to resolve
   its scopes.

From that perspective, it makes sense to hard-code TermTypes as an
"enum" of sorts. The one trade-off we make is that TermTypes can no
longer be accessed via the API as a resource list. You can still
aggregate on `term_type_id` via search to discover the list of
values programmatically, but I think ultimately, the way to solve
this is through good documentation.

All tests pass and the relevant artwork field still work well.