Refactor Category and Term to extend CategoryTerm
Both Category and Term now share the same table. The `category_terms`
view was replaced with a `category_terms` table. Likewise, Category
and Term now extend CategoryTerm. This is similar to how we handle
Asset and its child models. The crucial difference is that `assets`
is a search alias of its child resources, while `category-terms` is
the actual index, while `categories` and `terms` are search scopes,
defined in CategoryTerm.

Note that this approach necessitates defining fields in CategoryTerm,
not its children. Otherwise, they wouldn't be picked up when the
index is created – at least now with our current architecture.

On category-terms:

 - Added `subtype` - stored as id in db, but output as enum:

   classification, material, technique,
   style, subject, department, theme

On terms:

 - Removed `term_type_id`

On categories:

 - Removed `description`
 - Removed `is_in_nav`
 - Removed `sort`
 - Removed `type`

We can add these fields back if needed, but they are obsolete.

I kept `parent_id` b/c it offers a basis for building an ontology.