Add `technique_ids` to Artwork
Required to support filtering artworks by technique behind-the-scenes.
We aren't planning to add a "technique" filter to the sidebar, since
the data is pretty scarce, but technique terms are valid options for
the "quick facet" list, so it needs to exist in the code.

Required an update to the foundation to fix this error:

  SQLSTATE[23000]: Integrity constraint violation: 1052
  Column 'updated_at' in order clause is ambiguous

To update the search index, run this in tinker on the test server:

  \App\Models\Collections\Artwork::has('techniques')->searchable();

About 6797 artworks currently have `technique` terms.