Refactor source date logic in models
This commit makes it so that the fact of whether or not a model has
"source dates" is defined in one place - i.e. somewhere in its class
hierarchy. For example, almost all `Collections` models share the
`citi_created_at` and `citi_modified_at` fields. If we wanted to
modify the way these fields worked, we'd have to make the changes
to all the models that declare them in their `$dates` property.
Now, we can overwrite the `getDates()` method to affect all
child models at once.

We still need to make transformers check their model's properties
to determine whether to show dates. I'm also not sure whether to
put this logic into `getDates()` or `getCasts()`. This seems like
a good start, though. Not all tests pass; we'll fix that next.